CHET-284: harvest error flag from quickstart form into shared componnt
